+++ js/util/__tests__/extractProps.spec.ts
@@ -85,6 +85,15 @@ describe('Extracting Props with a configuration', () => {
     expect(extractProps('Element', obj1, config2)).toEqual('[a=foo];[c=true];');
   });
 
+  test('can handle if a prop has a circular ref', () => {
+    const myProp: { foo?: any } = {};
+    myProp.foo = myProp;
+    const obj2 = _.merge({}, obj1, { stateNode: { props: { c: myProp } } });
+    expect(extractProps('Element', obj2, config)).toEqual(
+      '[a=foo];[c.foo.foo.foo=object Object];'
+    );
+  });
+
   test("functions don't come through", () => {
     const obj2 = _.merge({}, obj1, {
       stateNode: {
@@ -130,11 +139,11 @@ describe('Extracting Props with a configuration', () => {
     );
   });
 
-  test('removes any brackets in a prop', () => {
+  test('removes any reserved characters in a prop', () => {
     const obj2 = _.merge({}, obj1, {
       stateNode: {
         props: {
-          a: 'bracket][bracket][bracket]',
+          a: 'bracket]@[|=#;;bracket][bracket]',
         },
       },
     });
@@ -142,6 +151,22 @@ describe('Extracting Props with a configuration', () => {
     expect(extractProps('Element', obj2, config)).toEqual(
       '[a=bracketbracketbracket];[c=true];'
     );
+  });
+
+  test('removes any props with keys containing reserved characters', () => {
+    const obj2 = _.merge({}, obj1, {
+      stateNode: {
+        props: {
+          'abc@;|=#': 'foo',
+        },
+      },
+    });
+
+    const config2 = _.merge({}, config, {
+      Element: { include: ['a', 'abc@;|=#'] },
+    });
+
+    expect(extractProps('Element', obj2, config2)).toEqual('[a=foo];');
   });
 
   test('extracts class configurations', () => {