+++ js/util/extractProps.ts
@@ -1,4 +1,9 @@
+import * as React from 'react';
 import { getCombinedInclusionList } from './combineConfigs';
+import {
+  containsReservedCharacter,
+  stripReservedCharacters,
+} from './reservedCharacters';
 
 const _ = require('lodash');
 const flatten = require('flat');
@@ -87,7 +92,16 @@ export const extractProps = (
     props = fiberNode.memoizedProps;
   }
 
-  const filteredProps = _.pick(props, inclusionList);
+  const filteredProps = _(props)
+    .pick(inclusionList)
+    .mapValues((prop: any) => {
+      if (React.isValidElement(prop)) {
+        // :TODO: (jmtaber129): Consider pulling information from the React element.
+        return 'React.element';
+      }
+      return prop;
+    })
+    .value();
 
   // KLUDGE: We want to capture the `key` property for list components that have it set,
   // but we can't simply add to the list of props captured for all components in
@@ -96,7 +110,7 @@ export const extractProps = (
   // prop for its own use; it is intended to be reserved for internal use. (HEAP-8473)
   const flattenedProps = Object.assign(
     { key: fiberNode.key },
-    flatten(filteredProps)
+    flatten(filteredProps, { maxDepth: 4 })
   );
 
   let propsString = '';
@@ -108,8 +122,14 @@ export const extractProps = (
       flattenedProps[key] !== undefined &&
       typeof flattenedProps[key] !== 'function'
     ) {
-      // Remove all brackets from string.
-      let prop = flattenedProps[key].toString().replace(/[\[\]]/g, '');
+      if (containsReservedCharacter(key)) {
+        console.warn(
+          `Prop key '${key}' contains reserved characters; ignoring.`
+        );
+        return;
+      }
+
+      const prop = stripReservedCharacters(flattenedProps[key].toString());
       propsString += `[${key}=${prop}];`;
     }
   });